New validate hook for the plugins
Add a new validate hook so plugins can validate the user input in their
own UI. This improves the usability since user errors are catched as
soon as possible. Otherwise weird things happen when you click on a
tab and you go back to the initial tab and your input is gone.