Support the mailto protocol in contact addresses
The email field in ContactPerson nodes should start with the
'mailto:' protocol, since it is an absolute URIs.

This change adds support to validate emails with and without
'mailto:', it validates emails without it because there are
federations using emails without the prefix, even if it is wrong.

This change also adds the prefix to an address if needed when
added through the graphical interface.